toolkit/content/tests/widgets/test_bug1319301.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Video controls test - bug 1319301</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div id="content">
  <video id="video" controls preload="auto"></video>
</div>

<pre id="test">
<script clas="testbody" type="application/javascript">
  const video = document.getElementById("video");
  const controlsSpacer = getAnonElementWithinVideoByAttribute(video, "anonid", "controlsSpacer");

  add_task(async function setup() {
    await new Promise(resolve => window.addEventListener("load", resolve));
    await SpecialPowers.pushPrefEnv({"set": [["media.cache_size", 40000]]});
  });

  add_task(async function play_fadeout() {
    isnot(controlsSpacer.getAttribute("fadeout"), "true", "controlsSpacer should not fadeout before playing");

    await new Promise(resolve => {
      video.addEventListener("canplaythrough", video.play);
      video.addEventListener("play", () => SimpleTest.executeSoon(resolve));
      video.src = "seek_with_sound.ogg";
    })

    is(controlsSpacer.getAttribute("fadeout"), "true", "controlsSpacer should fadeout once video starts playing");
  });
</script>
</pre>
</body>
</html>
